# openapi.yml
openapi: 3.1.0
info:
  title: Система бронирования мест в кафе
  version: 0.0.2
servers:
  - url: http://localhost:8000/
    description: Локальный сервер
...
paths:
...
  /actions:
    get:
      tags:
        - Акции
      summary: Получение списка акций
      description: Получение списка акций. Для администраторов и менеджеров - все акции (с возможностью выбора), для пользователей - только активные.
      operationId: get_all_dishes_actions_get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: show_all
          in: query
          required: false
          schema:
            type: boolean
            title: Показывать все акции?
            description: Показывать все акции или нет. По умолчанию показывает все акции
            default: false
          description: Показывать все акции или нет. По умолчанию показывает все акции
        - name: cafe_id
          in: query
          required: false
          schema:
            type: integer
            description: ID кафе, в котором показывать акции. Если не задано - показывает все акции во всех кафе
            title: Cafe Id
          description: ID кафе, в котором показывать акции. Если не задано - показывает все акции во всех кафе
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActionInfo'
                title: Response Get All Dishes Actions Get
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
    post:
      tags:
        - Акции
      summary: Создание новой акции
      description: Создает новую акцию. Только для администраторов и менеджеров.
      operationId: create_action_actions_post
      security:
        - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionCreate'
              description: Данные новой акции
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionInfo'
        '400':
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
  /actions/{action_id}:
    get:
      tags:
        - Акции
      summary: Получение информации об акции по ее ID
      description: Получение информации об акции по ее ID. Для администраторов и менеджеров - все акции, для пользователей - только активные.
      operationId: get_action_by_id_actions__action_id__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: action_id
          in: path
          required: true
          schema:
            type: integer
            description: ID акции
            title: Action Id
          description: ID акции
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionInfo'
        '400':
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '404':
          description: Данные не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
    patch:
      tags:
        - Акции
      summary: Обновление информации об акции по ее ID
      description: Обновление информации об акции по ее ID. Только для администраторов и менеджеров.
      operationId: path_dish_by_id_actions__action_id__patch
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: action_id
          in: path
          required: true
          schema:
            type: integer
            description: ID акции
            title: Action Id
          description: ID акции
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionUpdate'
              description: Обновленные данные акции
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionInfo'
        '400':
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '404':
          description: Данные не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
...
components:
  schemas:
    ActionCreate:
      properties:
        cafes_id:
          items:
            type: integer
          type: array
          title: Cafes Id
        description:
          type: string
          title: Description
        photo_id:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
          title: Photo Id
      type: object
      required:
        - cafes_id
        - description
        - photo_id
      title: ActionCreate
    ActionInfo:
      properties:
        id:
          type: integer
          title: Id
        cafes:
          items:
            $ref: '#/components/schemas/CafeShortInfo'
          type: array
          title: Cafes
        description:
          type: string
          title: Description
        photo_id:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
          title: Photo Id
        is_active:
          type: boolean
          title: Is Active
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
        - id
        - cafes
        - description
        - photo_id
        - is_active
        - created_at
        - updated_at
      title: ActionInfo
    ActionUpdate:
      properties:
        cafes_id:
          anyOf:
            - items:
                type: integer
              type: array
            - type: 'null'
          title: Cafes Id
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
        photo_id:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
          title: Photo Id
        is_active:
          anyOf:
            - type: boolean
            - type: 'null'
          title: Is Active
      type: object
      required:
        - cafes_id
        - description
        - photo_id
        - is_active
      title: ActionUpdate
...
    CafeShortInfo:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        address:
          type: string
          title: Address
        phone:
          type: string
          title: Phone
        description:
          type: string
          title: Description
        photo_id:
          type: string
          format: uuid
          title: Photo Id
      type: object
      required:
        - id
        - name
        - address
        - phone
        - description
        - photo_id
      title: CafeShortInfo
...
    CustomError:
      properties:
        code:
          type: integer
          title: Code
        message:
          type: string
          title: Message
      type: object
      required:
        - code
        - message
      title: CustomError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /auth/login